# faster d1() calc