On my computer, it's consistently 0.25 µs faster.

`3.84 µs ± 50.3 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)`

instead of 

`4.11 µs ± 15.5 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)`

And if you don't trust my math here's [wolfram](https://www.wolframalpha.com/input?i=%28log%28F%2FK%29+%2B+v%5E2+*+t+%2F+2%29+%2F+%28v+*+sqrt%28t%29%29)